Updated to a higher level of Knowledge
1. Rewrote my commands. Now the command and the listener is in same
class.
2. Registering my commands without plugin.yml
Using reflection in CommandMaper to get access to the commandmap and
registering the plugin.
Next step could be the implementation of configcommands, like in
plugin.yml but you'll be able to change command, usage, description and
aliases. For a proper alpha build of next version write me a
personalmessage.